+++ src/common/chunked/rfifo.h
@@ -50,7 +50,7 @@ struct fifo_chunk_buf {
 	if (p ## _flag == 0 || (p ## _flag == 2 && *p ## _dst_data_ptr == NULL)) { \
 		*p ## _dst_data_ptr = aMalloc(p ## _src_data_size); \
 		memcpy(*p ## _dst_data_ptr, p ## _src_data, p ## _src_data_size); \
-		*p ## _dst_data_size_ptr = p ## _src_data_size; \
+		*p ## _dst_data_size_ptr = (int)p ## _src_data_size; \
 	} else if (p ## _flag == 1 || p ## _flag == 2) { \
 		*p ## _dst_data_ptr = aRealloc(*p ## _dst_data_ptr, *p ## _dst_data_size_ptr + p ## _src_data_size); \
 		memcpy(*p ## _dst_data_ptr + *p ## _dst_data_size_ptr, p ## _src_data, p ## _src_data_size); \
